events: move isEventUnsupportedForFastPath to fast event emitter file
