+++ lib/internal/events/event_emitter.js
@@ -42,7 +42,7 @@ const {
   throwErrorOnMissingErrorHandler,
 } = require('internal/events/shared_internal_event_emitter');
 const { SlowEventEmitter } = require('internal/events/slow_event_emitter');
-const { FastEventEmitter } = require('internal/events/fast_event_emitter');
+const { FastEventEmitter, isEventUnsupportedForFastPath } = require('internal/events/fast_event_emitter');
 
 
 let EventEmitterAsyncResource;
@@ -213,32 +213,6 @@ ObjectDefineProperty(EventEmitter.prototype, kCapture, {
   writable: true,
   enumerable: false,
 });
-
-function isEventUnsupportedForFastPath(type) {
-  return (
-    // Not supporting newListener and removeListener events in fast path
-    // as they can add new listeners in the middle of the process
-    type === 'newListener' ||
-    type === 'removeListener' ||
-
-    // Not supporting errorMonitor event as it has special handling
-    type === kErrorMonitor ||
-
-    // Not supporting Object prototype keys as fast path object is not an object with null prototype
-    type === 'constructor' ||
-    type === '__defineGetter__' ||
-    type === '__defineSetter__' ||
-    type === 'hasOwnProperty' ||
-    type === '__lookupGetter__' ||
-    type === '__lookupSetter__' ||
-    type === 'isPrototypeOf' ||
-    type === 'propertyIsEnumerable' ||
-    type === 'toString' ||
-    type === 'valueOf' ||
-    type === '__proto__' ||
-    type === 'toLocaleString'
-  );
-}
 
 ObjectDefineProperties(EventEmitter.prototype, {
   [kImpl]: {
