+++ lib/internal/events/fast_event_emitter.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { kShapeMode } = require('internal/events/symbols');
+const { kShapeMode, kErrorMonitor } = require('internal/events/symbols');
 const {
   throwErrorOnMissingErrorHandler,
   addCatch,
@@ -9,10 +9,7 @@ const {
 /**
  * This class is optimized for the case where there is only a single listener for each event.
  * it supports kCapture
- * but does not support event types with the following names:
- * 1. 'newListener' and 'removeListener' as they can cause another event to be added
- *     and making the code less predictable thus harder to optimize
- * 2. kErrorMonitor as it has special handling
+ * but does not support event types that {@link isEventUnsupportedForFastPath} returns false for
  */
 function FastEventEmitter(eventEmitterTranslationLayer, _events) {
   this.eventEmitterTranslationLayer = eventEmitterTranslationLayer;
@@ -138,6 +135,35 @@ FastEventEmitter.prototype.removeAllListeners = function removeAllListeners(type
   return undefined;
 };
 
+
+function isEventUnsupportedForFastPath(type) {
+  return (
+    // Not supporting newListener and removeListener events in fast path
+    // as they can add new listeners in the middle of the process
+    type === 'newListener' ||
+    type === 'removeListener' ||
+
+    // Not supporting errorMonitor event as it has special handling
+    type === kErrorMonitor ||
+
+    // Not supporting Object prototype keys as fast path object is not an object with null prototype
+    type === 'constructor' ||
+    type === '__defineGetter__' ||
+    type === '__defineSetter__' ||
+    type === 'hasOwnProperty' ||
+    type === '__lookupGetter__' ||
+    type === '__lookupSetter__' ||
+    type === 'isPrototypeOf' ||
+    type === 'propertyIsEnumerable' ||
+    type === 'toString' ||
+    type === 'valueOf' ||
+    type === '__proto__' ||
+    type === 'toLocaleString'
+  );
+}
+
+
 module.exports = {
   FastEventEmitter,
+  isEventUnsupportedForFastPath,
 };
